pyproject.toml
[project]
name = "dropsort"
version = "0.1.0"
description = "Document classifier as an authenticated service (AIE39 Week 6)"
requires-python = ">=3.11,<3.12"
authors = [
    { name = "Fakih (MA)" },
    { name = "Nasser" },
    { name = "Saleh" },
    { name = "Shari" },
]
dependencies = [
    "fastapi>=0.115,<0.120",
    "fastapi-users[sqlalchemy]>=14,<15",
    "uvicorn[standard]>=0.32,<0.40",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.5,<3",
    "sqlalchemy>=2.0.30,<3",
    "asyncpg>=0.29,<1",
    "alembic>=1.13,<2",
    "casbin>=1.36,<2",
    "casbin-sqlalchemy-adapter>=1.4,<2",
    "fastapi-cache2[redis]>=0.2.1,<0.3",
    "redis>=4.2,<5",
    "rq>=2,<3",
    "hvac>=2.3,<3",
    "minio>=7.2,<8",
    "paramiko>=3.5,<4",
    "structlog>=24,<26",
    "httpx>=0.27,<0.30",
    "pillow>=12.2.0",
    "torch>=2.4,<3",
    "torchvision>=0.19,<1",
    "passlib[argon2]>=1.7,<2",
    "argon2-cffi>=23,<24",
    "fastapi-limiter>=0.1.6,<0.2",
    "email-validator>=2,<3",
    "psycopg2-binary>=2.9,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<2",
    "ruff>=0.7,<1",
    "mypy>=1.13,<2",
    "asgi-lifespan>=2,<3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
extend-exclude = [
    "notebook",
    "*.ipynb",
]